server/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
sqlalchemy==2.0.28
python-dotenv==1.0.1
python-multipart==0.0.9
Pillow==10.2.0
imagehash==4.3.1
ultralytics==8.3.30
pytest==7.3.2
